requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
Pillow==10.0.0
google-auth-oauthlib==1.1.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-api-python-client==2.99.0 